requirements.txt
alembic==1.14.0
altgraph==0.17.4
auto-py-to-exe==2.42.0
bcrypt==4.2.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distlib==0.3.8
filelock==3.13.1
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==23.2
pycparser==2.21
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
efficientnet-pytorch==0.7.1
Flask==3.0.3
Flask-Cors==5.0.0
setuptools==69.1.0
virtualenv==20.25.1
Werkzeug==3.1.3
